WIP - Progress on authorized routes
