+++ Atlas/frontend/src/components/NotFound/NotFound.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NotFound = () => {
+  return <div>NOT FOUND</div>;
+}
+
+export default NotFound;
