+++ Atlas/frontend/src/routes/Profile/Profile.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+import './style.css';
+
+export default class LoginRoute extends Component {
+  render() {
+    return (
+      <div>PROFILE</div>
+    );
+  }
+};
